+++ test/td_dd_web/controllers/data_structure_controller_test.exs
@@ -9,7 +9,6 @@ defmodule TdDdWeb.DataStructureControllerTest do
   alias TdDd.DataStructures.RelationTypes
   alias TdDd.Lineage.GraphData
   alias TdDd.Permissions.MockPermissionResolver
-  alias TdDdWeb.ApiServices.MockTdAuditService
   alias TdDdWeb.ApiServices.MockTdAuthService
 
   import Routes
@@ -44,7 +43,6 @@ defmodule TdDdWeb.DataStructureControllerTest do
 
   setup_all do
     start_supervised(MockTdAuthService)
-    start_supervised(MockTdAuditService)
     start_supervised(MockPermissionResolver)
     start_supervised(PathCache)
     start_supervised(GraphData)